Refactor: useKeywordInput 분리
